Simplify state management and improve config structure
- Restructure Config to match config.yaml format directly
- Remove wrapper classes (MonitorConfig, TerminalConfig, TriggerManager)
- Move agent.rs to agent/mod.rs for cleaner structure
- Add comprehensive config tests (13 new tests, 130% increase)
- Consolidate helper functions and remove unnecessary abstractions
- Move code to appropriate modules (capture to rule/, filters to trigger/)
