+++ src/config/mod.rs
@@ -1,64 +1,57 @@
-pub mod agents;
-pub mod loader;
-pub mod rule;
-pub mod terminal;
-pub mod trigger;
-pub mod types;
-pub mod web_ui;
+pub mod agents_config;
+pub mod helper;
+pub mod rules_config;
+pub mod triggers_config;
+pub mod web_ui_config;
 
-use crate::config::loader::load_config;
-use crate::config::rule::Rule;
-use crate::config::trigger::TriggerManager;
+use crate::config::agents_config::AgentsConfig;
+use crate::config::rules_config::Rule;
+use crate::config::triggers_config::Trigger;
+use crate::config::web_ui_config::WebUIConfig;
 use anyhow::Result;
-use std::sync::Arc;
-use tokio::sync::RwLock;
+use serde::Deserialize;
+use std::convert::TryFrom;
 
+/// Main configuration structure matching config.yaml format
+#[derive(Debug, Deserialize, Clone, Default)]
 pub struct Config {
-    trigger_manager: TriggerManager,
-    rules: Arc<RwLock<Vec<Rule>>>,
-    // Monitor configuration
-    monitor_config: loader::MonitorConfig,
+    #[serde(default)]
+    pub web_ui: WebUIConfig,
+    #[serde(default)]
+    pub agents: AgentsConfig,
 }
 
 impl Config {
-    pub async fn new(config_path: &str) -> Result<Self> {
-        // Load initial configuration (triggers, rules, and monitor config)
-        let (initial_triggers, initial_rules, monitor_config) =
-            load_config(std::path::Path::new(config_path))?;
-        let trigger_manager = TriggerManager::new(initial_triggers);
-        let rules = Arc::new(RwLock::new(initial_rules));
+    /// Load configuration from YAML file
+    pub fn from_file(config_path: &str) -> Result<Self> {
+        let content = std::fs::read_to_string(config_path)
+            .map_err(|e| anyhow::anyhow!("Failed to read config file {}: {}", config_path, e))?;
 
-        Ok(Config {
-            trigger_manager,
-            rules,
-            monitor_config,
-        })
+        let config: Config = serde_yml::from_str(&content)
+            .map_err(|e| anyhow::anyhow!("Failed to parse YAML config: {}", e))?;
+
+        Ok(config)
     }
 
-    /// Get monitor configuration
-    pub fn get_monitor_config(&self) -> &loader::MonitorConfig {
-        &self.monitor_config
+
+    /// Parse triggers from config
+    pub fn parse_triggers(&self) -> Result<Vec<Trigger>> {
+        let mut triggers = Vec::new();
+        for trigger_config in &self.agents.triggers {
+            let trigger = Trigger::try_from(trigger_config.clone())?;
+            triggers.push(trigger);
+        }
+        Ok(triggers)
     }
 
-    /// Get trigger manager
-    pub fn get_trigger_manager(&self) -> &TriggerManager {
-        &self.trigger_manager
-    }
-
-    /// Get parsed rules directly
-    pub fn get_rules(&self) -> Vec<Rule> {
-        self.rules.try_read().map(|r| r.clone()).unwrap_or_default()
-    }
-}
-
-impl std::fmt::Debug for Config {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Config")
-            .field(
-                "rules_count",
-                &self.rules.try_read().map(|r| r.len()).unwrap_or(0),
-            )
-            .finish()
+    /// Parse rules from config
+    pub fn parse_rules(&self) -> Result<Vec<Rule>> {
+        let mut rules = Vec::new();
+        for rule_config in &self.agents.rules {
+            let rule = Rule::try_from(rule_config.clone())?;
+            rules.push(rule);
+        }
+        Ok(rules)
     }
 }
 
@@ -66,24 +59,20 @@ impl std::fmt::Debug for Config {
 mod tests {
     use super::*;
 
-    #[tokio::test]
-    async fn test_get_rules() {
-        // Use existing diff-timeout-demo config
-        let config_path = "examples/diff-timeout-demo/config.yaml";
-        let config = Config::new(config_path).await.unwrap();
+    #[test]
+    fn test_config_from_file() {
+        // Test loading a basic config file
+        let config_path = "examples/basic/config.yaml";
+        let config = Config::from_file(config_path).unwrap();
 
-        // Test that we can get rules directly
-        let rules = config.get_rules();
+        // Test web_ui section
+        assert!(config.web_ui.enabled);
+        assert_eq!(config.web_ui.host, "localhost");
+        assert_eq!(config.web_ui.base_port, 9990);
 
-        // Should have some rules (the diff-timeout-demo config has rules)
-        assert!(!rules.is_empty(), "Should have rules from config file");
-
-        // Test that we can call it multiple times
-        let rules2 = config.get_rules();
-        assert_eq!(
-            rules.len(),
-            rules2.len(),
-            "Should get same number of rules each time"
-        );
+        // Test agents section
+        assert_eq!(config.agents.pool, 1);
+        assert!(!config.agents.triggers.is_empty());
+        assert!(!config.agents.rules.is_empty());
     }
 }
