+++ src/web_server/tests.rs
@@ -1,13 +1,13 @@
 use super::server::WebServer;
 use crate::agent::Agent;
-use crate::config::loader::MonitorConfig;
-use crate::config::web_ui::WebUIConfig;
+use crate::config::Config;
+use crate::config::web_ui_config::WebUIConfig;
 use crate::web_ui::assets::AssetCache;
 
 #[tokio::test]
 async fn test_web_server_creation() {
-    let monitor_config = MonitorConfig::default();
-    let agent = Agent::from_monitor_config(0, &monitor_config)
+    let config = Config::default();
+    let agent = Agent::from_config(0, &config)
         .await
         .unwrap();
     let web_server = WebServer::new(8080, "localhost".to_string(), agent);
@@ -25,11 +25,11 @@ fn test_web_ui_config_defaults() {
 }
 
 #[test]
-fn test_monitor_config_with_web_ui() {
-    let config = MonitorConfig::default();
+fn test_config_with_web_ui() {
+    let config = Config::default();
 
-    assert_eq!(config.get_web_ui_port(), 9990);
-    assert_eq!(config.get_agent_pool_size(), 1);
+    assert_eq!(config.web_ui.base_port, 9990);
+    assert_eq!(config.agents.pool, 1);
     assert!(config.web_ui.enabled);
     assert_eq!(config.web_ui.host, "localhost");
     assert_eq!(config.web_ui.base_port, 9990);
